+++ src/api/index.ts
@@ -1,7 +1,15 @@
 import { Adsapi } from './lib/api';
 export default Adsapi;
 
-export type { IADSApiBootstrapResponse, IUserData } from './lib/accounts/types';
+export type { IADSApiBootstrapResponse, IUserData } from '../_api/accounts/types';
+export type { SolrField, SolrSort, SolrSortDirection, SolrSortField } from '../_api/models';
+export type { IADSApiReferenceParams, IADSApiReferenceResponse } from '../_api/reference/types';
+export type {
+  IADSApiSearchErrorResponse,
+  IADSApiSearchParams,
+  IADSApiSearchResponse,
+  IDocsEntity,
+} from '../_api/search/types';
 export type { ExportApiFormat, IExportApiParams, IExportApiResponse, isExportApiFormat } from './lib/export/types';
 export type { IADSApiGraphicsParams, IADSApiGraphicsResponse } from './lib/graphics/types';
 export type {
@@ -10,11 +18,3 @@ export type {
   IADSApiMetricsResponse,
   ReadsHistogramType,
 } from './lib/metrics/types';
-export type { SolrField, SolrSort, SolrSortDirection, SolrSortField } from './lib/models';
-export type { IADSApiReferenceParams, IADSApiReferenceResponse } from './lib/reference/types';
-export type {
-  IADSApiSearchErrorResponse,
-  IADSApiSearchParams,
-  IADSApiSearchResponse,
-  IDocsEntity,
-} from './lib/search/types';
