+++ src/api/lib/graphics/graphics.ts
@@ -1,7 +1,7 @@
 import { IADSApiGraphicsParams, IADSApiGraphicsResponse } from '@api';
 import { AxiosError, AxiosRequestConfig } from 'axios';
 import { err, ok, Result } from 'neverthrow';
-import { ApiTargets } from '../models';
+import { ApiTargets } from '../../../_api/models';
 import { Service } from '../service';
 export class GraphicsService extends Service {
   async query(params: IADSApiGraphicsParams): Promise<Result<IADSApiGraphicsResponse, Error | AxiosError>> {
