+++ src/api/lib/metrics/metrics.ts
@@ -1,7 +1,7 @@
 import { IADSApiMetricsParams, IADSApiMetricsResponse } from '@api';
 import { AxiosError, AxiosRequestConfig } from 'axios';
 import { err, ok, Result } from 'neverthrow';
-import { ApiTargets } from '../models';
+import { ApiTargets } from '../../../_api/models';
 import { Service } from '../service';
 
 export class MetricsService extends Service {
