--- src/api/lib/reference/index.ts
@@ -1,1 +1,0 @@
-export * from './reference';
