+++ src/api/lib/vault/types.ts
@@ -1,4 +1,4 @@
-import { IADSApiSearchParams } from '../search/types';
+import { IADSApiSearchParams } from '../../../_api/search/types';
 
 export interface IADSApiVaultParams extends Partial<IADSApiSearchParams> {
   bigquery: IADSApiSearchParams['q'];
