+++ src/api/lib/vault/vault.ts
@@ -1,7 +1,7 @@
 import { AxiosError, AxiosRequestConfig } from 'axios';
 import { err, ok, Result } from 'neverthrow';
 import qs from 'qs';
-import { ApiTargets } from '../models';
+import { ApiTargets } from '../../../_api/models';
 import { Service } from '../service';
 import { IADSApiVaultParams, IADSApiVaultResponse } from './types';
