+++ src/components/AbstractSources/linkGenerator.ts
@@ -1,5 +1,5 @@
 import { IDocsEntity } from '@api';
-import { Esources } from '@api/lib/search/types';
+import { Esources } from '@_api/search/types';
 import { compose, descend, is, map, prop, sortBy, sortWith } from 'ramda';
 import { DEFAULT_ORDERING, GATEWAY_BASE_URL, LINK_TYPES } from './model';
 import { getOpenUrl } from './openUrlGenerator';
