+++ src/components/Metatags/Metatags.tsx
@@ -1,6 +1,6 @@
 import { IDocsEntity } from '@api';
-import { Esources } from '@api/lib/search/types';
 import { getFomattedNumericPubdate } from '@utils';
+import { Esources } from '@_api/search/types';
 import getConfig from 'next/config';
 import { ReactElement } from 'react';
