+++ src/mocks/handlers.ts
@@ -1,8 +1,8 @@
 import { IADSApiGraphicsParams, IADSApiSearchResponse, IExportApiParams } from '@api';
 import { ApiTargets } from '@api/lib/models';
-import { Esources } from '@api/lib/search/types';
 import faker from '@faker-js/faker';
 import { IADSApiMetricsParams } from '@_api/metrics';
+import { Esources } from '@_api/search/types';
 import { rest } from 'msw';
 import qs from 'qs';
 import { map, range, slice } from 'ramda';
